Sorry it too so long to respond.
After looking up a bunch of stuff online I decided to implement the music list two .xml files, one for the main display and another for the formatting of each item in the list.

The only thing I see that's left to implement for the MusicListActivity is the on click listener to play the music.